Avoid adding shader multiple times (#1679)
* Avoid adding shader multiple times

Check shader signature on Java side before requesting a new shader from C++

* set u_right from camera mask

instead of from render mask

* Make u_right between 0 and 1

* Fix handling of u_right

* Avoid adding shader multiple times

Check shader signature on Java side before requesting a new shader from C++

* set u_right from camera mask

instead of from render mask

* Make u_right between 0 and 1

* Fix handling of u_right

* Fix merge problem in GVRShader

and remove unused bindShader variant

* clear RenderData dirty, call bindShaders when lights change

* Added copy constructor for GVRMaterial

This makes it a lot easier to change the shader associated with a material

* Fix bug with removing lights

* added NO_LOGGING #define to suppress logging

Setting this flag is commented out in gvr_log.h. If NO_LOGGING is defined, none of the C++ LOG calls do anything.

+++ GVRf/Framework/framework/src/main/jni/engine/renderer/renderer.h
@@ -74,7 +74,7 @@ struct ShaderUniformsPerObject {
     glm::mat4   u_mvp_[2];      // ModelViewProjection matrix
     glm::mat4   u_mv_it;        // inverse transpose of ModelView
     glm::mat4   u_mv_it_[2];    // inverse transpose of ModelView
-    int         u_right;        // 1 = right eye, 0 = left
+    float       u_right;        // 1 = right eye, 0 = left
 };
 
 struct RenderState {
